backend/requirements/base.txt

djangorestframework
psycopg2-binary
psycopg2
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
django-jazzmin
python-dotenv